refactor: Moved canvas elements to a subfolder
