+++ src/colorizer/canvas/elements/index.ts
@@ -1,0 +1,13 @@
+export { defaultFooterOptions, getFooterRenderer } from "./footer";
+export { defaultHeaderOptions, getHeaderRenderer } from "./header";
+export { defaultInsetBoxOptions } from "./insetBox";
+export { defaultLegendOptions } from "./legend";
+export { defaultScaleBarOptions } from "./scalebar";
+export { defaultTimestampOptions } from "./timestamp";
+
+export type { FooterOptions, FooterParams } from "./footer";
+export type { HeaderOptions, HeaderParams } from "./header";
+export type { InsetBoxOptions } from "./insetBox";
+export type { LegendOptions, LegendParams } from "./legend";
+export type { ScaleBarOptions, ScaleBarParams } from "./scalebar";
+export type { TimestampOptions, TimestampParams } from "./timestamp";
