+++ src/colorizer/canvas/index.ts
@@ -1,0 +1,3 @@
+export type { BaseRenderParams, ContainerOptions, FontStyleOptions, RenderInfo } from "./types";
+
+export { configureCanvasText, getPixelRatio, renderCanvasText } from "./utils";
